pyproject.toml
[project]
name = "sports-card-ai-agent"
version = "0.1.0"
description = "AI-powered multi-agent system for analyzing and trading sports cards"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }

# Core dependencies
dependencies = [
    # Web & UI
    "streamlit>=1.30.0",
    "streamlit-authenticator>=0.3.0",

    # AI & Agents
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "langchain-community>=0.3.0",

    # Data & Models
    "pydantic>=2.0.0",
    "pandas>=2.2.0",
    "sqlalchemy>=2.0.0",

    # HTTP & APIs
    "httpx>=0.27.0",
    "requests>=2.31.0",

    # Sports APIs
    "nba-api>=1.1.14",
    "nhl-api-py>=1.0.0",
    "MLB-StatsAPI>=1.7.0",

    # Web Scraping
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "fake-useragent>=1.5.0",

    # Sentiment Analysis
    "textblob>=0.17.1",
    "vaderSentiment>=3.3.2",

    # Database
    "psycopg2-binary>=2.9.0",

    # Production
    "gunicorn>=21.2.0",
    "python-dotenv>=1.0.0",
    "alembic>=1.13.0",
]

# Optional dependencies for specific features
[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "httpx>=0.27.0",
]
production = [
    "redis>=4.5.0",
    "celery>=5.3.0",
    "sentry-sdk>=1.38.0",
    "flask>=2.3.0",
]

# Tool configurations
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "E402"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
